+++ lib/facter/nagios_redis.rb
@@ -1,4 +1,10 @@
 # Create custom nagios_redis fact
-if File.exists?('/usr/bin/redis-server')
-  Facter.add('nagios_redis') { setcode { true } }
+binaries_redis = [
+  '/usr/bin/redis-server',
+  '/usr/bin/valkey-server',
+]
+binaries_redis.each do |filename|
+  if File.exists?(filename)
+    Facter.add('nagios_redis') { setcode { true } }
+  end
 end